+++ packages/person/src/person-card/event.ts
@@ -1,0 +1,17 @@
+import { AzureIdOrUpnObj } from './task';
+import { CardData } from './types';
+
+export class RequestResolvePersonCardEvent extends CustomEvent<
+  AzureIdOrUpnObj & { result?: CardData | Promise<CardData> }
+> {
+  static readonly eventName = 'request-resolve-person-card';
+  constructor(detail: AzureIdOrUpnObj, options: Omit<CustomEventInit, 'detail'> = { bubbles: true, composed: true }) {
+    super(RequestResolvePersonCardEvent.eventName, { detail, ...options });
+  }
+}
+
+declare global {
+  interface ElementEventMap {
+    [RequestResolvePersonCardEvent.eventName]: RequestResolvePersonCardEvent;
+  }
+}